requirements.txt
pillow-heif>=0.10.0
Pillow>=9.0.0
